# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShelfOps API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/incidents/{incidentId}": {
      "get": {
        "summary": "Get a visible incident",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "incidentId",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Incident"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "correlationId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "correlationId"
                      ]
                    }
                  ],
                  "unevaluatedProperties": false
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1api~1v1/get/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1api~1v1~1stores~1{storeId}~1locations~1{locationId}~1configuration/post/responses/401"
          },
          "403": {
            "description": "Default Response",
            "headers": {
              "X-Correlation-Id": {
                "schema": {
                  "type": "string"
                },
                "description": "Server-generated identifier for support correlation."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message",
                    "correlationId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "forbidden"
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "correlationId": {
                      "description": "Server-generated identifier for support correlation.",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/paths/~1api~1v1~1stores~1{storeId}~1locations~1{locationId}~1configuration/post/responses/404"
          },
          "503": {
            "$ref": "#/paths/~1api~1v1/get/responses/503"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "shelfops_session"
      }
    },
    "schemas": {
      "Incident": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storeId": {
            "type": "string",
            "format": "uuid"
          },
          "sectorId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "type": "string",
            "description": "Category key from the organization catalog."
          },
          "severity": {
            "type": "string"
          },
          "categoryProvisional": {
            "type": "boolean"
          },
          "severityProvisional": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "reporterId": {
            "type": "string",
            "format": "uuid"
          },
          "assigneeId": {
            "type": "string",
            "format": "uuid"
          },
          "ownershipGap": {
            "type": "boolean"
          },
          "state": {
            "type": "string",
            "enum": [
              "open",
              "classified",
              "in-progress",
              "blocked",
              "resolved"
            ]
          },
          "reopenCount": {
            "type": "integer"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "storeId",
          "sectorId",
          "locationId",
          "category",
          "severity",
          "categoryProvisional",
          "severityProvisional",
          "title",
          "description",
          "occurredAt",
          "reporterId",
          "ownershipGap",
          "state",
          "reopenCount",
          "version",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  }
}